#867 Add usageCount to Words
- chore(refactor) Moved `usageCount` from Word to Content.
- Added `usageCount` to REST API
