+++ pom.xml
@@ -9,7 +9,7 @@
     <properties>
     	<java.version>1.8</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <model.version>2.0.9</model.version> <!-- https://github.com/elimu-ai/model/releases -->
+        <model.version>2.0.10</model.version> <!-- https://github.com/elimu-ai/model/releases -->
         <hibernate.version>4.3.11.Final</hibernate.version> <!-- https://repo.maven.apache.org/maven2/org/hibernate/ -->
         <jetty.version>8.1.16.v20140903</jetty.version> <!-- https://repo.maven.apache.org/maven2/org/mortbay/jetty/ -->
         <spring.version>4.2.9.RELEASE</spring.version> <!-- https://repo.maven.apache.org/maven2/org/springframework/ -->
