+++ src/main/java/ai/elimu/model/content/Allophone.java
@@ -16,7 +16,6 @@ import ai.elimu.model.enums.content.allophone.SoundType;
 import ai.elimu.model.enums.content.allophone.VowelFrontness;
 import ai.elimu.model.enums.content.allophone.VowelHeight;
 import ai.elimu.model.enums.content.allophone.VowelLength;
-import ai.elimu.tasks.AllophoneUsageCountScheduler;
 
 /**
  * Speech sound
@@ -93,12 +92,6 @@ public class Allophone extends Content {
      */
     @Enumerated(EnumType.STRING)
     private ConsonantVoicing consonantVoicing;
-    
-    /**
-     * This count is automatically extracted from {@link StoryBook}s, and is regularly updated by 
-     * the {@link AllophoneUsageCountScheduler}.
-     */
-    private int usageCount;
 
     public String getValueIpa() {
         return valueIpa;
@@ -194,13 +187,5 @@ public class Allophone extends Content {
 
     public void setConsonantVoicing(ConsonantVoicing consonantVoicing) {
         this.consonantVoicing = consonantVoicing;
-    }
-
-    public int getUsageCount() {
-        return usageCount;
-    }
-
-    public void setUsageCount(int usageCount) {
-        this.usageCount = usageCount;
     }
 }
